Update README.md
ASP.NET Core Web API CRUD with Angular is a web application project that provides a basic implementation of a RESTful API using ASP.NET Core on the back end and Angular on the front end. The application allows users to perform CRUD (Create, Read, Update, Delete) operations on a list of cards.